+++ SmallBeachFarm/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Updated for Harmony upgrade in SMAPI 3.12.0.
+
 ## 1.9.1
 Released 24 July 2021 for SMAPI 3.9.5 or later.